airflow/requirements.txt
pandas==1.3.3
findspark
delta-spark==2.2.0
pydantic==1.10.10
boto3==1.16.46
pendulum==2.1.2
connexion==2.13.0
hdfs==2.6.0
minio==7.1.0
poyo==0.5.0
pytest==6.2.5
pytz==2021.1
requests==2.26.0
apache-airflow[postgres,password]==2.1.4
joblib==1.0.1
pytest-spark==0.6.0
psycopg2-binary==2.9.1
SQLAlchemy==1.3.24
Flask-SQLAlchemy==2.5.1
configparser==5.0.2
Jinja2==2.11.3
Werkzeug==1.0.1
